create account

+++ controllers/usersController.js
@@ -1,35 +1,45 @@
 import asyncHandler from "../middleware/asyncHandler.js";
-import User from "../models/User.js";
 import ErrorResponse from "../utils/ErrorResponse.js";
+import User from "../models/User.js";
+import Account from "../models/Account.js";
+
 // @desc    Get all Users
 // @route   GET /api/v1/users
 // @access  Public
 export const getUsers = asyncHandler(async (req, res, next) => {
-  const users = await User.find({});
-  res.status(200).json({
-    success: true,
-    data: users,
-  });
+  const users = await User.find();
+  res.status(200).json({ data: users });
 });
 
 // @desc    Create a User
 // @route   POST /api/v1/users
 // @access  Private
 export const createUser = asyncHandler(async (req, res, next) => {
-  console.log(req.body);
-  const user = await User.create(req.body);
+  let user = await User.create(req.body);
+
+  const account = await Account.create({ owner: user._id });
+
+  user = await User.findByIdAndUpdate(
+    user._id,
+    {
+      $push: { accounts: account._id },
+    },
+    { new: true }
+  );
 
   res.status(200).json({
     success: true,
-    data: `User ${user.email} was created`,
+    data: { user, account },
   });
 });
 
-// @desc    Get a single user
+// @desc    Get a single User
 // @route   GET /api/v1/users/:id
 // @access  Public
 export const getUser = asyncHandler(async (req, res, next) => {
-  const user = await User.findById(req.params.id);
+  const user = await User.findById(req.params.id).populate({
+    path: "accounts",
+  });
 
   if (!user) {
     return next(
@@ -43,13 +53,13 @@ export const getUser = asyncHandler(async (req, res, next) => {
   });
 });
 
-// @desc    Get users by
-// @route   GET /api/v1/users/getUserBy
-// @access  Public
-export const getUserBy = asyncHandler(async (req, res, next) => {
+//@desc Get a User by Query
+//@route GET /api/v1/users/getUserByQuery
+//@access
+export const getUserByQuery = asyncHandler(async (req, res, next) => {
   const query = req.query;
-  if (!query.hasOwnProperty("email")) {
-    return next(new ErrorResponse("Params can only be Email/userID"), 401);
+  if (!query.hasOwnProperty("email") && !query.hasOwnProperty("passportID")) {
+    return next(new ErrorResponse("Params can only be Email/passportID"), 401);
   }
   const user = await User.find(query);
   if (!user) {
@@ -62,41 +72,12 @@ export const getUserBy = asyncHandler(async (req, res, next) => {
   });
 });
 
-// @desc    update a user balance
-// @route   GET /api/v1/users/updateBalance/:id
-export const updateBalance = asyncHandler(async (req, res, next) => {
-  const userId = req.params.id;
-  const { credit = 0, cash = 0 } = req.body;
-
-  const { credit: prevCredit, cash: prevCash } = await User.findById(userId);
-  if (!prevCredit || !prevCash) {
-    new ErrorResponse(`User that ends with '${userId}' was not found`, 404);
-  }
-
-  // runValidators will not work with $inc
-  const user = await User.findOneAndUpdate(
-    { _id: userId },
-    { credit: credit + prevCredit, cash: cash + prevCash },
-    { new: true, runValidators: true }
-  );
-  if (!user) {
-    return next(
-      new ErrorResponse(`User that ends with '${userId}' was not found`, 404)
-    );
-  }
-
-  res.status(200).json({
-    success: true,
-    data: `Balance was Changed`,
-  });
-});
-
-// @desc    Delete a user
+// @desc    Delete a User
 // @route   DELETE /api/v1/users/:id
 // @access  Private
 export const deleteUser = asyncHandler(async (req, res, next) => {
   const user = await User.findById(req.params.id);
-  const userEmail = user.email;
+
   if (!user) {
     return next(
       new ErrorResponse(
@@ -106,10 +87,17 @@ export const deleteUser = asyncHandler(async (req, res, next) => {
     );
   }
 
+  const accounts = await Account.deleteMany({ owner: user._id });
+  if (!accounts.acknowledged) {
+    return next(
+      new ErrorResponse(`Error occurred while deleting user accounts`, 400)
+    );
+  }
+
   user.deleteOne();
 
   res.status(200).json({
     success: true,
-    data: `User ${userEmail} was deleted`,
+    data: "User and accounts removed",
   });
 });
